backend/requirements.txt
backoff==2.2.1
redis==4.4.0
elastic-transport==8.4.0
elasticsearch[async]==8.6.1
fastapi==0.87.0
gunicorn==20.1.0
orjson==3.8.2
pydantic==1.10.2
uvicorn==0.20.0
uvloop==0.17.0
python-dotenv==0.21.0
pyjwt==2.6.0
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-exporter-jaeger==1.15.0
